+++ packages/ui/src/components/index.ts
@@ -40,6 +40,8 @@ export {
   Snackbar,
   Radio,
   Select,
+  InputLabel,
+  Fade,
 } from '@mui/material';
 
 export type {
@@ -82,6 +84,8 @@ export type {
   GridProps,
   SnackbarProps,
   SelectProps,
+  InputLabelProps,
+  FadeProps,
 } from '@mui/material';
 
 export { LoadingButton, type LoadingButtonProps } from '@mui/lab';