+++ src/components/Transfer/TransferAsset.tsx
@@ -1,122 +1,34 @@
+import { useEffect } from 'react';
 import { observer } from 'mobx-react-lite';
-import { Stack, Typography, MenuItem, styled, AmountInput, LoadingButton } from '@cere-wallet/ui';
-import { Button, FormControl, TextField } from '@cere/ui';
-import * as yup from 'yup';
-import { SubmitHandler, useForm } from 'react-hook-form';
-import { yupResolver } from '@hookform/resolvers/yup';
-import { AssetSelectItem } from '~/components';
-import { useAssetStore, useNetworkStore } from '~/hooks';
-import { useCallback, useEffect, useState } from 'react';
-import { Asset } from '~/stores';
-import { Divider } from '@mui/material';
+import { Controller } from 'react-hook-form';
 import { useNavigate } from 'react-router-dom';
-import { useTransactionStore } from '~/hooks/useTransactionStore';
-import { BigNumber } from 'bignumber.js';
+import { Stack, Typography, AmountInput, LoadingButton, Button, TextField, Alert } from '@cere-wallet/ui';
 
-const ContentSelect = styled(TextField)(() => ({
-  '& .MuiSelect-select': {
-    padding: 0,
-  },
-  '& .MuiInputBase-root': {
-    minHeight: 57,
-  },
-}));
-
-const Summary = styled(Stack)(({ theme }) => ({
-  backgroundColor: theme.palette.neutral.light,
-  borderRadius: 16,
-}));
-
-const validationSchema = yup
-  .object({
-    asset: yup.string().required('Asset is required field'),
-    address: yup
-      .string()
-      .required('Address is required field')
-      .matches(/^0x[a-fA-F0-9]{40}$/g, 'Should be ethereum address format'),
-    amount: yup
-      .string()
-      .required('Amount is required field')
-      .matches(/^\d*\.?\d*$/, 'Should be a number'),
-  })
-  .required();
+import { useAssetStore, useTransactionStore } from '~/hooks';
+import { AssetSelect } from '../AssetSelect';
+import { useTransferForm } from './useTransferForm';
 
 const TransferAsset = () => {
   const navigate = useNavigate();
-
   const { list } = useAssetStore();
-  const network = useNetworkStore();
-  const { transferErc20 } = useTransactionStore();
-  const [selectedAsset, setSelectedAsset] = useState<Asset | undefined>();
-  const [total, setTotal] = useState<string>('0');
+  const { transfer } = useTransactionStore();
+  const assets = list.filter((asset) => asset.ticker !== 'CERE'); // TODO: Remove when CERE transfer implementation
 
   const {
+    control,
     register,
     handleSubmit,
-    setError,
-    getValues: getFormValues,
+    resetField,
+    watch,
     formState: { errors, isSubmitting },
-  } = useForm({
-    resolver: yupResolver(validationSchema),
-    mode: 'onSubmit',
-    defaultValues: {
-      asset: undefined,
-      address: '0x9317382d69804b22f2b5d1779ecbf62c8c11aa67',
-      amount: '0',
-    },
-  });
+  } = useTransferForm({ assets });
 
-  const updateTotal = useCallback(() => {
-    const amountValue = getFormValues('amount');
-    if (!+amountValue || !selectedAsset) {
-      setTotal(`${network?.fee} ${network?.network?.ticker}`);
-      return;
-    }
-
-    const tax = new BigNumber(network?.fee);
-    const amount = new BigNumber(getFormValues('amount'));
-    if (network.network?.ticker === selectedAsset?.ticker) {
-      const total = tax.plus(amount);
-      setTotal(`${total.toString()} ${network.network?.ticker}`);
-    } else {
-      setTotal(`${amount.toString()} ${selectedAsset?.ticker} + ${network.fee} ${network.network?.ticker}`);
-    }
-  }, [selectedAsset, network, getFormValues]);
-
-  const amountValidate = (): boolean => {
-    const value: string = getFormValues('amount');
-
-    if (selectedAsset) {
-      const maxBalance = new BigNumber(selectedAsset?.balance || '0');
-      const amount = new BigNumber(value);
-      if (amount.gt(maxBalance)) {
-        setError('amount', {
-          message: `this field cannot be greater than balance (${selectedAsset?.ticker} balance is ${
-            selectedAsset?.balance || 0
-          })`,
-        });
-        return false;
-      }
-    }
-
-    updateTotal();
-
-    return true;
-  };
-
-  const navigateBackHandler = useCallback(() => {
-    navigate(-1);
-  }, [navigate]);
-
-  const onSubmit: SubmitHandler<any> = async ({ asset, address, amount }) => {
-    if (asset && address && amount && amountValidate()) {
-      await transferErc20(asset, address, amount); // TODO integrate transaction page here
-    }
-  };
+  const ticker = watch('asset');
+  const selectedAsset = assets.find((asset) => asset.ticker === ticker);
 
   useEffect(() => {
-    updateTotal();
-  }, [selectedAsset, updateTotal]);
+    resetField('amount');
+  }, [resetField, ticker]);
 
   return (
     <Stack
@@ -126,87 +38,52 @@ const TransferAsset = () => {
       component="form"
       noValidate
       autoComplete="off"
-      onSubmit={handleSubmit(onSubmit)}
+      onSubmit={handleSubmit(async ({ asset, address, amount }) => {
+        try {
+          await transfer(asset, address, amount);
+
+          resetField('amount');
+        } catch (error) {
+          console.warn(error);
+        }
+      })}
     >
       <Stack>
         <Typography variant="body2">Select asset</Typography>
-        <FormControl>
-          <ContentSelect
-            select
-            {...register('asset')}
-            error={!!errors?.asset?.message}
-            helperText={errors.asset?.message}
-            required
-            autoFocus
-            name="asset"
-            onChange={({ target }) => setSelectedAsset(list.find((asset) => asset.ticker === target.value))}
-            sx={{ padding: 0 }}
-          >
-            {list.map((asset, index) => (
-              <MenuItem key={`${asset.ticker}-${index}`} value={asset.ticker} sx={{ borderRadius: 0, height: 57 }}>
-                <AssetSelectItem
-                  coin={asset.ticker}
-                  name={asset.displayName}
-                  network={asset.network}
-                  balance={asset.balance || 0}
-                />
-              </MenuItem>
-            ))}
-          </ContentSelect>
-        </FormControl>
+        <Controller name="asset" control={control} render={({ field }) => <AssetSelect assets={assets} {...field} />} />
       </Stack>
       <Stack>
         <Typography variant="body2">Transfer To</Typography>
-        <FormControl>
-          <TextField
-            {...register('address')}
-            error={!!errors?.address?.message}
-            helperText={errors.address?.message}
-            placeholder="Recipient’s ETH address"
-            name="address"
-            variant="outlined"
-          />
-        </FormControl>
+
+        <TextField
+          {...register('address')}
+          required
+          error={!!errors.address?.message}
+          helperText={errors.address?.message}
+          placeholder="Recipient’s address"
+        />
       </Stack>
+
       <Stack>
         <Typography variant="body2">Amount</Typography>
-        <FormControl>
-          <AmountInput
-            maxValue={selectedAsset?.balance}
-            {...register('amount')}
-            error={!!errors?.amount?.message}
-            helperText={errors.amount?.message}
-            onKeyUp={amountValidate}
-            required
-            name="amount"
-            variant="outlined"
-          />
-          {selectedAsset && (
-            <Typography variant="body2" color="text.secondary">
-              Available balance: {selectedAsset?.balance} {selectedAsset?.ticker}
-            </Typography>
-          )}
-        </FormControl>
+
+        <AmountInput
+          {...register('amount')}
+          required
+          placeholder="0"
+          maxValue={selectedAsset?.balance?.toString()}
+          error={!!errors.amount?.message}
+          helperText={
+            errors.amount?.message ||
+            (selectedAsset?.balance && `Available balance: ${selectedAsset.balance} ${selectedAsset.displayName}`)
+          }
+        />
       </Stack>
-      <Summary spacing={2} alignItems="stretch" padding={2}>
-        <Stack direction="row" justifyContent="space-between">
-          <Typography variant="subtitle1" color="text.secondary" fontWeight="regular">
-            Network fee
-          </Typography>
-          <Typography variant="subtitle1">
-            {network.fee?.toString()} {network.network?.ticker}
-          </Typography>
-        </Stack>
-        <Divider />
-        <Stack direction="row" justifyContent="space-between">
-          <Typography variant="subtitle1" color="text.secondary" fontWeight="regular">
-            Total cost
-          </Typography>
-          <Typography variant="subtitle1">{total}</Typography>
-        </Stack>
-      </Summary>
-      <Stack direction="row" spacing={2}>
-        <Button fullWidth variant="outlined" onClick={() => navigateBackHandler()}>
+
+      {isSubmitting && <Alert severity="info">Your transfer transaction is being processed</Alert>}
+
+      <Stack direction="row" spacing={2} paddingTop={2}>
+        <Button fullWidth disabled={isSubmitting} variant="outlined" onClick={() => navigate(-1)}>
           Cancel
         </Button>
         <LoadingButton fullWidth loading={isSubmitting} variant="contained" type="submit">