+++ src/routes/Transfer/TransferCollectibles.tsx
@@ -1,10 +1,13 @@
 import { observer } from 'mobx-react-lite';
-import { Stack } from '@cere-wallet/ui';
+import { ComingSoon, Stack } from '@cere-wallet/ui';
 
 const TransferCollectibles = () => {
   return (
     <Stack alignItems="stretch" spacing={2}>
-      TransferCollectibles
+      <ComingSoon
+        title="Collectibles transfer"
+        description="Soon you will be able to transfer your collectibles. Stay tuned for updates."
+      />
     </Stack>
   );
 };